refactor: writeType0Packet 모듈화
connect.ts에서 정의되었던 writePacket 함수를 writeType0Packet으로 바꾸고
packet.ts에 모듈화 하였다.
