+++ mediaServer/core/rtmp/packet.ts
@@ -1,3 +1,5 @@
+import net from 'net';
+
 type RtmpChunckBasicHeader = {
   chunkType: number;
   chunkStreamId: number;
@@ -136,4 +138,14 @@ function separateRtmpChunk(data: Buffer): RtmpChunck[] {
   return rtmpChunks;
 }
 
-export { separateRtmpChunk, RtmpChunck };
+function writeType0Packet(socket: net.Socket, streamId: number, messageTypeId: number, payload: Buffer) {
+  const basicHeader = Buffer.alloc(1);
+  basicHeader.writeIntLE(streamId, 0, 1);
+  const messageHeader = Buffer.alloc(11);
+  messageHeader.writeIntBE(payload.length, 3, 3); // message length 설정
+  messageHeader.writeIntBE(messageTypeId, 6, 1); // message type id 설정
+
+  socket.write(Buffer.concat([basicHeader, messageHeader, payload]));
+}
+
+export { separateRtmpChunk, writeType0Packet, RtmpChunck };
